Add GCP Project ID information to fss namespaces
This information is used by logging operator to send logs to google
cloud logs in the team's project as a part of the new team logs setup.

The namespace label is used when all logs are sent to team logs, while the annotation is used by naiserator to populate the `GOOGLE_PROJECT_ID` environment variable which is used when sending only selected logs to team logs.

Co-authored-by: Terje Sannum <terje.sannum@nav.no>